[GH-19] Clean up some types
