+++ src/components/TimeTable/helper.ts
@@ -1,7 +1,43 @@
-import { IDay, ISlot } from 'types/TimeTable'
-
 export const timeSlot = generateTimeSlot(8, 20)
 export const dayInWeek = ['จ', 'อ', 'พ', 'พฤ', 'ศ', 'ส', 'อา']
+
+export enum DAY_IN_WEEK {
+  MONDAY = 1,
+  TUESDAY,
+  WEDNESDAY,
+  THURSDAY,
+  FRIDAY,
+  SATURDAY,
+  SUNDAY,
+}
+
+export interface IDay {
+  dayInWeek: DAY_IN_WEEK
+  subjectList: ISubject[]
+}
+
+export enum SUBJECT_TYPE {
+  LECTURE = 'LECTURE',
+  LAB = 'LAB',
+}
+
+export interface ISubject {
+  id: string
+  code: string
+  name: string
+  section: number
+  startSlot: number
+  endSlot: number
+  type: SUBJECT_TYPE
+  isEditing?: boolean
+}
+
+export interface ISlot {
+  slotSpan: number
+  startSlot: number
+  endSlot: number
+  subjectList: ISubject[]
+}
 
 /**
  * Input 7 days data from backend and this hooks
